+++ ai/common/cells_array.py
@@ -1,5 +1,4 @@
 
-
 class CellsArray:
     '''
     a 2d array representing all the cells
