+++ ai/common/test_board_status.py
@@ -1,6 +1,6 @@
 from .helpers import p
-from .Board import Board
-from .BoardStatus import BoardStatus
+from .board import Board
+from .board_status import BoardStatus
 
 def test_board_state_black_won():
     layout1 = p("X------X",
